sprint 15

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -2,8 +2,9 @@ package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.repository.BookingRepository;
 import ru.practicum.shareit.exception.NotFoundException;
@@ -15,16 +16,12 @@ import ru.practicum.shareit.item.repository.ItemRepository;
 import ru.practicum.shareit.user.service.UserService;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
+import java.util.*;
 
-@Transactional(readOnly = true)
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class ItemServiceImpl implements ItemService {
-
 
     private final ItemRepository itemRepository;
     private final BookingRepository bookingRepository;
@@ -38,20 +35,21 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<Item> getAllItemsByOwnerId(Long ownerId) {
-        List<Item> itemList = itemRepository.findByOwnerId(ownerId);
-        itemList.sort(Comparator.comparing(Item::getId));
+    public List<Item> getAllItemsByOwnerId(Long ownerId, Integer from, Integer size) {
+        Pageable pageable = PageRequest.of(from / size, size);
+        List<Item> itemList = itemRepository.findByOwnerId(ownerId, pageable);
+        if (itemList.size() > 1) {
+            itemList.sort(Comparator.comparing(Item::getId));
+        }
         return itemList;
     }
 
-    @Transactional
     @Override
     public Item createItem(Long ownerId, Item item) {
         item.setOwner(userService.getUserById(ownerId));
         return itemRepository.save(item);
     }
 
-    @Transactional
     @Override
     public Item updateItem(Long ownerId, Item item) {
         Item updateItem = itemRepository.findById(item.getId())
@@ -72,14 +70,14 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<Item> searchItemsByTextInNameAndDescription(String text) {
+    public List<Item> searchItemsByTextInNameAndDescription(String text, Integer from, Integer size) {
         if (text.isBlank()) {
             return Collections.emptyList();
         }
-        return itemRepository.searchItemsByTextInNameAndDescription(text);
+        Pageable pageable = PageRequest.of(from / size, size);
+        return itemRepository.searchItemsByTextInNameAndDescription(text, pageable);
     }
 
-    @Transactional
     @Override
     public Comment createComment(Long userId, Long itemId, Comment comment) {
         Booking booking = bookingRepository.findCompletedBooking(userId, itemId, LocalDateTime.now());
@@ -97,7 +95,16 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public void deleteItem(Long itemId) {
+    public List<Item> getItemsByRequestId(Long requestId) {
+        return itemRepository.findItemsByRequestId(requestId);
+    }
+
+    @Override
+    public void deleteItemById(Long ownerId, Long itemId) {
+        Item item = getItemById(itemId);
+        if (!item.getOwner().getId().equals(ownerId)) {
+            throw new ValidationException("Удалить вещь может только её владелец");
+        }
         itemRepository.deleteById(itemId);
     }
 }
